fix(backup): listdir Windows error path used %lu in lua_pushfstring
lua_pushfstring understands only a fixed set of conversions (%s/%d/%f/
%p/%c/%%) - never %lu. When FindFirstFileA/FindNextFileA failed, the
(nil, err) branch fed it "(error %lu)", which does not return an error
string but RAISES "invalid option '%l' to 'lua_pushfstring'", breaking
the documented (nil, err) contract and crashing whatever called listdir.

The backup engine collects scripts/cfg, which does not exist in a fresh
tree: on POSIX listdir returned (nil, "...errno 2") cleanly (opendir path
uses %d) and the engine skipped the missing dir, but on Windows the same
missing-dir case hit the %lu path and raised, wedging +backup with a
script error and the smoke reply timeout (Windows-only CI failure).

Cast the DWORD error to int (Win32 error codes are small, positive) so
lua_pushfstring builds the message and listdir honours its (nil, err)
contract on Windows, matching POSIX.

Part of #480

+++ hub/hub.c
@@ -259,8 +259,11 @@ static int listdir(lua_State *L)
   if (h == INVALID_HANDLE_VALUE)
   {
     lua_pushnil(L);
-    lua_pushfstring(L, "listdir: cannot open '%s' (error %lu)",
-                    path, (unsigned long)GetLastError());
+    /* lua_pushfstring understands only %d/%s/%f/%p/%c/%% - NOT %lu; feeding it
+     * %lu raises "invalid option '%l'" and breaks the (nil, err) contract, so
+     * the DWORD error is cast to int (codes are small, positive). */
+    lua_pushfstring(L, "listdir: cannot open '%s' (error %d)",
+                    path, (int)GetLastError());
     return 2;
   }
   lua_newtable(L);
@@ -283,8 +286,8 @@ static int listdir(lua_State *L)
   {
     lua_pop(L, 1);
     lua_pushnil(L);
-    lua_pushfstring(L, "listdir: enumeration failed for '%s' (error %lu)",
-                    path, (unsigned long)ferr);
+    lua_pushfstring(L, "listdir: enumeration failed for '%s' (error %d)",
+                    path, (int)ferr);
     return 2;
   }
   return 1;
